pyproject.toml
[project]
name = "claude-agent-experiment"
version = "0.1.0"
description = "Code to experiment with the claude agentic capabilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.20",
    "composio>=0.10.8",
    "composio-claude-agent-sdk>=0.10.8",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.11",
    "httpx>=0.27.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
]
